fix: moves the key function call to inside of the GoogleMap composable feat: update to Places SDK 4.0.0.  remove places-ktx dependency chore: clean up log statements and unused code
