+++ places-compose-demo/src/main/java/com/google/android/libraries/places/compose/demo/data/repositories/CountriesRepository.kt
@@ -17,6 +17,10 @@ import com.google.android.libraries.places.compose.demo.data.models.Country
 
 class CountriesRepository {
     companion object {
+
+        /**
+         * Returns a flag emoji for the given country code.
+         */
         private fun getFlagEmoji(countryCode: String): String {
             val firstLetter = Character.codePointAt(countryCode, 0) - 0x41 + 0x1F1E6
             val secondLetter = Character.codePointAt(countryCode, 1) - 0x41 + 0x1F1E6
