+++ places-compose-demo/src/main/java/com/google/android/libraries/places/compose/demo/data/repositories/MergedLocationRepository.kt
@@ -2,7 +2,6 @@ package com.google.android.libraries.places.compose.demo.data.repositories
 
 import android.Manifest
 import android.annotation.SuppressLint
-import android.util.Log
 import androidx.annotation.RequiresPermission
 import com.google.android.gms.maps.model.LatLng
 import kotlinx.coroutines.CoroutineScope
@@ -13,7 +12,6 @@ import kotlinx.coroutines.flow.WhileSubscribed
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.mapNotNull
-import kotlinx.coroutines.flow.merge
 import kotlinx.coroutines.flow.shareIn
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.seconds
@@ -24,6 +22,9 @@ data class CompositeLocation(
     val isMockLocation: Boolean = true,
 )
 
+/**
+ * A repository class that allows switching between the actual device location and a mock location.
+ */
 class MergedLocationRepository
 @Inject
 constructor(
@@ -33,31 +34,11 @@ constructor(
 ) {
   private val _useMockLocation = MutableStateFlow(true)
 
-  val mergedLocation = merge(
-    locationRepository.latestLocation.mapNotNull { latLng ->
-      latLng?.let {
-        CompositeLocation(
-          latLng = it,
-          label = "Current Location",
-          isMockLocation = false,
-        )
-      }
-    },
-    mockLocationRepository.location.map {
-      CompositeLocation(
-        latLng = it.latLng,
-        label = it.label,
-        isMockLocation = true,
-      )
-    },
-  )
-
   @SuppressLint("MissingPermission")
   @ExperimentalCoroutinesApi
   @RequiresPermission(allOf = [Manifest.permission.ACCESS_COARSE_LOCATION, Manifest.permission.ACCESS_FINE_LOCATION])
   val location = _useMockLocation.flatMapLatest { useMockLocation ->
     if (useMockLocation) {
-      Log.d("MergedLocationRepository", "Emitting mock location")
       mockLocationRepository.location.map {
         CompositeLocation(
             latLng = it.latLng,
@@ -66,7 +47,6 @@ constructor(
         )
       }
     } else {
-      Log.d("MergedLocationRepository", "Emitting system location")
       locationRepository.latestLocation.mapNotNull { latLng ->
         latLng?.let {
           CompositeLocation(
